simplify switch
